apply task
